refactor: extract cli/which handler (was backup/whichDumpTool)
The tool finder isn't backup-specific anymore — it's used by Azure and AWS
CLI discovery too. Move it to its own cliHandlers module and rename the IPC
method to cli/which. Also drop a pointless length>0 guard around the PATH
concat (a leading empty spread + filter(Boolean) is a no-op).

Updates all callers: BaseCommandClient (backup), CliPathPicker (Azure/AWS),
the integration test setup, and the RCE security spec.
