+++ apps/studio/src-commercial/backend/handlers/backupHandlers.ts
@@ -1,26 +1,13 @@
 import { Command } from "@/lib/db/models";
 import { spawn } from "child_process";
-import path from "path";
 import { state } from "@/handlers/handlerState";
-import platformInfo from "@/common/platform_info";
 
 export const errorMessages = {
   nonZero: 'Command returned non-zero exit code'
 }
 
-// Directories to search for backup tools in addition to the inherited PATH.
-// macOS GUI apps don't inherit the shell's PATH, so Homebrew's bin dirs are
-// missing — add them explicitly so brew-installed tools (pg_dump, mysqldump,
-// az, ...) are discovered. Apple Silicon installs to /opt/homebrew/bin, Intel
-// to /usr/local/bin. `which` returns the stable symlink there, which keeps
-// pointing at the current version after a `brew upgrade`.
-export const extraToolSearchDirs: string[] = platformInfo.isMac
-  ? ['/opt/homebrew/bin', '/usr/local/bin']
-  : [];
-
 export interface IBackupHandlers {
   'backup/runCommand': ({ command, sId }: { command: Command, sId: string }) => Promise<void>,
-  'backup/whichDumpTool': ({ toolName }: { toolName: string }) => Promise<string>,
   'backup/cancelCommand': ({ sId }: { sId: string }) => Promise<boolean>
 }
 
@@ -107,44 +94,6 @@ export const BackupHandlers: IBackupHandlers = {
         })
       })
     }
-  },
-  'backup/whichDumpTool': async function({ toolName }: { toolName: string }) {
-    const command = platformInfo.isWindows ? 'where' : 'which';
-
-    // Pass the environment explicitly so we can prepend the Homebrew bin dirs
-    // that a macOS GUI app would otherwise be missing from its PATH.
-    const env = { ...process.env };
-    if (extraToolSearchDirs.length > 0) {
-      env.PATH = [...extraToolSearchDirs, env.PATH].filter(Boolean).join(path.delimiter);
-    }
-
-    return new Promise((resolve, reject) => {
-      const proc = spawn(command, [toolName], { shell: false, env });
-
-      let stdout = '';
-      let stderr = '';
-
-      proc.stdout.on('data', (chunk) => {
-        stdout += chunk.toString();
-      });
-
-      proc.stderr.on('data', (chunk) => {
-        stderr += chunk.toString();
-      });
-
-      proc.on('error', (err) => {
-        reject(err);
-      });
-
-      proc.on('close', (code) => {
-        if (code === 0) {
-          const path = stdout.trim().split('\n')[0]; // pick first result
-          resolve(path);
-        } else {
-          reject(`whichTool failed (code ${code})\nSTDERR: ${stderr}\nSTDOUT: ${stdout}`);
-        }
-      });
-    });
   },
   'backup/cancelCommand': async function({ sId }: { sId: string }) {
     if (state(sId).backupProc) {
