+++ apps/studio/src-commercial/backend/handlers/handlers.ts
@@ -8,6 +8,7 @@ import { IConnectionHandlers } from "./connHandlers";
 import { IExportHandlers } from "./exportHandlers";
 import { IImportHandlers } from "./importHandlers";
 import { IBackupHandlers } from "./backupHandlers";
+import { ICliHandlers } from "./cliHandlers";
 import { IEnumHandlers } from "./enumHandlers";
 import { IAwsHandlers } from "./awsHandlers";
 
@@ -18,6 +19,7 @@ export interface Handlers
     IImportHandlers,
     IExportHandlers,
     IBackupHandlers,
+    ICliHandlers,
     IFileHandlers,
     IEnumHandlers,
     ITempHandlers,
