+++ apps/studio/tests/unit/security/backupHandlers.exploit.spec.ts
@@ -37,6 +37,7 @@ jest.mock("@/handlers/handlerState", () => ({
 }));
 
 import { BackupHandlers } from "@commercial/backend/handlers/backupHandlers";
+import { CliHandlers } from "@commercial/backend/handlers/cliHandlers";
 
 function pwnPath(label: string): string {
   return path.join(
@@ -117,13 +118,13 @@ describe("backupHandlers RCE — live touch demonstrations", () => {
   );
 
   skipOnWindows(
-    "backup/whichDumpTool executes attacker shell payload injected via toolName (shell:true)",
+    "cli/which executes attacker shell payload injected via toolName (shell:true)",
     async () => {
       const pwnFile = pwnPath("which-toolname");
       expect(fs.existsSync(pwnFile)).toBe(false);
 
       try {
-        await BackupHandlers["backup/whichDumpTool"]({
+        await CliHandlers["cli/which"]({
           // `which` runs, fails on the first arg, then `; touch FILE` runs.
           toolName: `bogus-tool; touch '${pwnFile}'; #`,
         }).catch(() => {});
